\name{BreakpointsCallCommands}
\alias{BreakpointsCallCommands}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Command Generator for using BreakpointsCall }
\description{
Generates a Unix command (one per batch) to calculate Breakpoints using BreakpointsCall R program
}
\usage{
BreakpointsCallCommands(RDfile_RepSample,max_bin_size=1000,\ 
BPfile,Rbatch_folder="")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{RDfile_RepSample}{ Read depth file of a representative sample in the analysis batch (any sample should work)  }
  \item{max_bin_size}{ Bin size for number of probe regions from which MAD is estimated (default=1000 should be enough for most cases) }
  \item{BPfile}{ Breakpoints file generated by BreakpointsCall }
  \item{Rbatch_folder}{	[optional] Rbatch folder bundled with CoNVex for batch execution. Rbatch_folder="" (default, works mostly) assumes that it's in R's library folder. You may copy it to somewhere else and modify this explicitly }
}
\details{
BreakpointsCallCommands.
}
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  This can be easily used together with batch queuing (e.g., bsub) programs. This function usually one Unix command for all samples in the batch.
}
\seealso{ 
BreakpointsCall, Breakpoints, cut
}
\examples{
> gam_commands = BreakpointsCallCommands(RDfile_RepSample,\ 
max_bin_size=1000,BPfile,Rbatch_folder="")
> gam_commands[1]
[1] "R --vanilla --slave --args '/home/pv1/UK10K/ProbeRD_UK10K12345.dat,\ 
/home/pv1/UK10K/MOPD_Features.txt,\ 
/home/pv1/UK10K/L2R/GAM_MOPD5095427.dat,\ 
/home/pv1/UK10K/ProbeRD_MOPD5095427.dat,\ 
/home/pv1/UK10K/MOPD_Breakpoints.txt' < \ 
/home/pv1/R/x86_64-linux/2.11/CoNVex/Rbatch/GAMCorrectionPerSample.R"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ breakpoints, median absolute deviation }
