\name{GAMCorrection}
\alias{GAMCorrection}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Systematic Error Correction of log2 Ratio }
\description{
GAMCorrection corrects log2 ratio for systematic errors caused by GC content, deltaG and Tm. GC content is the proportion of GC in probe regions that causes systematic variation in read depth. 
deltaG is the free energy of hybridization between biotinylated RNA probes and sheared genomic DNA. Tm is the melting temperature of hybridization. deltaG and Tm are thermodynamic features that estimate the systematic fluctuation
of read depth caused by variation in hybridization efficiency between RNA probes and genomic DNA. This function depends on 'mgcv' package.
}
\usage{
GAMCorrection(L2Rfile,features_file,RDfile,BPfile,doNorm=0,gc_only=0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{L2Rfile}{ File containing log2 ratio of a specific sample }
  \item{features_file}{ A subset of probe regions file with rowcount, Chr regions, GC, deltaG and Tm as columns }
  \item{RDfile}{ Read depth file of a specific sample }
  \item{BPfile}{	Breakpoints file generated by Breakpoints / BreakpointsCall }
  \item{doNorm}{ doNorm=0 (default) does not do any normalization (recommended). Other values normalize (Z-score) the log2 ratio. In future, this can be extended for better normalization if required. }
  \item{gc_only}{ gc_only=0 (default) Use GC only for correction (=1 enables GC only). }
}
\details{
GAMCorrection function corrects systematic errors in log2 ratio (of depth/sample-median) for each probe region. 
This function uses the gam() function of the 'mgcv' package and its multivariate cubic splines model to correct the log2 ratio using GC, deltaG and Tm. 
deltaG and Tm are inter-related thermodynamic features So, gam() considers them as causing synergetic effect and uses s(deltaG,Tm). GAMCorrectionPerSample.R calls 
it for all samples. GAMCorrectionCommands generates Unix commands for batch execution. 
}
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  This can be easily used together with batch queuing (e.g., bsub) programs. This function usually one Unix command 
per sample in the analysis.
}
\references{ 
Wood, S.N. (2011) Fast stable restricted maximum likelihood and marginal likelihood estimation of
semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1):3-36
}
\seealso{ 
GAMCorrectionCommands, GAMCorrectionPerSample \cr
mgcv:gam - \url{http://people.bath.ac.uk/sw283/mgcv/} \cr
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ log2 ratio, GC content, deltaG, Tm, target capture efficiency, systematic error correction }
