\name{GAMCorrectionCommands}
\alias{GAMCorrectionCommands}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Unix Command Generator for using GAMCorrectionPerSample }
\description{
Generates Unix command(s) (e.g., one per batch) to calculate corrected Log2 Ratio and error-weighted Scores
}
\usage{
GAMCorrectionCommands(features_file,L2Rfiles,GAMfiles,RDfiles,BPfile,output_folder,sample_ids,gc_only=1,Rbatch_folder="")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{features_file}{ A subset of probe regions file with rowcount, Chr regions, GC, deltaG and Tm as columns }
  \item{L2Rfiles}{ Vector of files containing log2 ratio generated by SampleLogRatio }
  \item{GAMfiles}{ Vector of file names to save the corrected log2 ratio and error-weighted scores }
  \item{RDfiles}{ Vector of read depth files }
  \item{BPfile}{	Breakpoints file generated by BreakpointsCall }
  \item{output_folder}{	Output folder for read depth files }
  \item{sample_ids}{ Vector of Sample IDs }
  \item{gc_only}{ Use GC only for correction, avoid deltaG and TM }
  \item{Rbatch_folder}{ [optional] Rbatch folder bundled with CoNVex for batch execution. Rbatch_folder="" (default, works mostly) assumes that it's in R's library folder. You may copy it to somewhere else and modify this explicitly }
}
\details{
GAMCorrection function corrects systematic errors in log2 ratio (of depth/sample-median) for each probe region. GAMCorrectionPerSample.R calls it for all samples. 
GAMCorrectionCommands (this function) generates Unix commands for batch execution. This function depends on 'mgcv' package.
}
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  This can be easily used together with batch queuing (e.g., bsub) programs. This function usually one Unix command per sample in the analysis.
}
\references{ 
Wood, S.N. (2011) Fast stable restricted maximum likelihood and marginal likelihood estimation of
semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1):3-36
}
\seealso{ 
GAMCorrection,
GAMCorrectionPerSample,\cr
mgcv:gaml - \url{http://people.bath.ac.uk/sw283/mgcv/} \cr
}
\examples{
> gam_commands = GAMCorrectionCommands(features_file,L2Rfiles,\ 
GAMfiles,RDfiles,BPfile,output_folder,\ 
sample_ids,Rbatch_folder="");
> gam_commands[1]
[1] "R --vanilla --slave --args '/home/pv1/UK10K/ProbeRD_UK10K12345.dat,\ 
/home/pv1/UK10K/MOPD_Features.txt,\ 
/home/pv1/UK10K/L2R/GAM_MOPD5095427.dat,\ 
/home/pv1/UK10K/ProbeRD_MOPD5095427.dat,\ 
/home/pv1/UK10K/MOPD_Breakpoints.txt' < \ 
/home/pv1/R/x86_64-linux/2.11/CoNVex/Rbatch/GAMCorrectionPerSample.R"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ log2 ratio, error-weighted scores }
