\name{MeansMadsCommands}
\alias{MeansMadsCommands}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Mean log2 ratio and MAD of mean log2 ratios for each CNV call  }
\description{
Mean log2 ratio and MAD of mean log2 ratios for each CNV call.
}
\usage{
MeansMadsCommands(CNVfiles,sample_ids,cor_matrix,sample_info_file,regions_file, features_file, all_samples=0, Rbatch_folder="",output_folder=".")
}
\arguments{
  \item{CNVfiles}{	Vector of output files to store CNV calls }
  \item{sample_ids}{ Vector of Sample IDs }
  \item{cor_matrix}{ Correlation matrix }
  \item{sample_info_file}{ Tab-separated file auto-generated by SamplePrepInfo() function - required for the whole analysis }
  \item{regions_file}{ Probe regions file }
  \item{features_file}{ A subset of probe regions file with rowcount, Chr regions, GC, deltaG and Tm as columns }
  \item{all_samples}{ [default = 0] Consider all samples in the analysis to calculate means and mads (no need for cor_matrix file; suitable for small projects) }
  \item{Rbatch_folder}{	[optional] Rbatch folder bundled with CoNVex for batch execution. Rbatch_folder="" (default, works mostly) assumes that it's in R's library folder. You may copy it to somewhere else and modify this explicitly }
  \item{output_folder}{	[optional] Output folder }
}
\details{
MeansMadsCommands calls MeansMadsCall
}
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  # Output file name is auto-generated (ending with: MeansMads.txt)
}
\seealso{ 
MeansMadsCall is a batch execution script
}
\examples{
> #
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ MeansMadsCommands }
