\name{MultiPanelPlot}
\alias{MultiPanelPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Zoomed-in visualisation of each CNV calls  }
\description{
Plots a CNV call and other information in multiple panels.
}
\usage{
MultiPanelPlot(allX,allXscore,Fd_rep,sample_id,known_rstr, gene_info,misc_str="",mother_id="",father_id="")
}
\arguments{
  \item{allX}{ Contains log2 ratio of sample(s) to be plotted and a set of background samples }
  \item{allXscore}{ Contains error-weighted scores of sample(s) to be plotted and a set of background samples }
  \item{Fd_rep}{ Single CNV call - a row in the data frame generated by GetCNVCalls() }
  \item{sample_id}{ Sample ID of this CNV call - scores of this sample must be present in allX and allXscore }
  \item{known_rstr}{ List of known CNV calls in this region }
  \item{gene_info}{ A data frame containing genes in one column (chr, start, end in first 3 columns, and 'gene_symbol' column) }
  \item{misc_str}{ Misc short string that goes in the legend  }
  \item{mother_id}{ Mother's sample id if present in allX  }
  \item{father_id}{ Father's sample id if present in allXscore }
}
\details{
MultiPanelPlot has 3 panels: (1) log2 ratio of CNV call; (2) Error-weighted score of CNV call; (3) Information panel
}
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  # Output contains 1 plot in the current directory. You may like to use setwd("your_folder") to set the current directory. File names of the png files
are determined automatically. Rows in gene_file (if present) should match the rows in regions_file used for CNV calling. 
Columns of the file include: rowcount, chr, start, end, gene_symbols, transcripts. (you can substituet gene symbols and transcripts 
with any other relevant information instead.
}
\seealso{ 
PlotCNVStats
}
\examples{
> #
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ MultiPanelPlot }
