\name{SampleInfoPrepInteractive}
\alias{SampleInfoPrepInteractive}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Input Data Validation before Starting the Analysis }
\description{
This function does basic validation with the input and creates output folders/files for further CNV analysis. 
This is an important first step to start the analysis. The output file (sample info file) created by this function 
has columns/info in a defined format. Returns the same info as a data frame.
}
\usage{
SampleInfoPrepInteractive(sample_ids,gender,bamfiles,RDfiles,\ 
L2Rfiles,GAMfiles,sample_info_file,output_folder,overwrite=FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sample_ids}{ Vector of Sample IDs }
  \item{gender}{ Vector of Sample gender }
  \item{regions_file}{ Probe regions file}
  \item{bamfiles}{ Vector of BAM files - rows should match sample_ids }
  \item{RDfiles}{ Vector of output read depth files }
  \item{L2Rfiles}{ Vector of file names in which uncorrected log2 ratio generated by SampleLogRatio will be stored - full path required }
  \item{GAMfiles}{ Vector of file names in which corrected log2 ratio and MAD-weighted scores generated by GAMCorrection will be stored - full path required}
  \item{sample_info_file}{ Output file name to store the all required sample information for further analysis}
  \item{output_folder}{	[optional] Output folder for read depth and other files }
  \item{overwrite}{ [optional; default=FALSE] By default, sample info file is not overwritten as a precaution. overwrite=TRUE overwrites it }
}
\details{
 #  Requires input info as shown in function options\cr
 #  Creates output_folder if it doesn't exist\cr
 #  Checks whether the BAM files exist \cr
 #  Checks gender - only 'F', 'M' and 'U' are allowed\cr
 #  Checks whether sample_info_file exists already - overwrite=TRUE (default: FALSE) overwrites it without warning\cr
 #  Saves the sample_info_file in the specified output_folder with all required columns\cr
 #  Output saved in sample_info_file is also returned as a data frame. You can keep it in an R interactive session for reference.
}
\references{ Agilent's ADM2 algorithm }
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  This step is required for further analysis. Based on your expertise, you may create the sample_info_file yourself through other ways. 
You'll save a lot of time before the analysis by making sure all input and output files are right place by using this simple function.
}
\seealso{ 
CoNVex documentation
}
\examples{
require(CoNVex)
baminfoRD = read.table("/path/to/SampleIDs_Gender_BAMfileLocations.txt", sep="\t")

# Required info
# regions - string / given with CoNVex package
regions_file = paste(getLibPath(),"/CoNVex/extdata/SureSelect_50Mb.txt",sep=""); 
sample_ids = as.character(baminfoRD[,3]); 
bamfiles = as.character(baminfoRD[,4]);  
gender = as.character(baminfoRD[,2]); 
gender[gender=='Male'] = 'M'; 
gender[gender=='Female'] = 'F';
output_folder = "/home/pv1/UK10K/";
RDfiles = paste(output_folder,"/ProbeRD_",sample_ids,".dat",sep="") 
L2Rfiles = paste(output_folder,"/L2R/L2R_",sample_ids,".dat",sep="") 
GAMfiles = paste(output_folder,"/L2R/GAM_",sample_ids,".dat",sep="") 
sample_info_file = "/home/pv1/UK10K/UK10K_SampleInfo.txt"

# parameters of the SW algorithm and CNVfiles
p=3; swt_del=5; swt_dup=5; dv=0.5; 
sw_exec = paste(getLibPath(),"/CoNVex/exec/swa_lin64",sep=""); # SW-Array execution binary
swa_folder = paste(output_folder,"/CNVcalls_p",p,"_t",t,"_new",sep=""); # CNV calls will be stored in this folder:
CNVfiles = paste(swa_folder,"/CoNVex_",sample_ids,"_p",p,"_tdel",swt_del,"_tdup",swt_dup,"_dv",dv,"_.txt",sep="") # one file per sample

# These files are not required for this function, but REQUIRED LATER
BPfile =  "/home/pv1/UK10K/MOPD_Breakpoints.txt"
centromere_regions_file = paste(getLibPath(),"/CoNVex/extdata/gaps_table_hg19.txt",sep="")
features_file = "/home/pv1/UK10K/MOPD_Features.txt"

# Save all vectors in a fixed format
baminfoALL = SampleInfoPrepInteractive(sample_id=sample_ids, gender=gender, bamfiles=bamfiles, RDfiles=RDfiles, L2Rfiles=L2Rfiles, GAMfiles=GAMfiles, CNVfiles=CNVfiles, sample_info_file=sample_info_file, output_folder=output_folder, overwrite = FALSE);
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ read depth, reads, read count, depth, picard }
