\name{SampleLogRatioCallCommands}
\alias{SampleLogRatioCallCommands}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Unix Command Generator for using SampleLogRatioCall }
\description{
Generates Unix command(s) (e.g., one per batch) to calculate Sample Log2 Ratio 
}
\usage{
SampleLogRatioCallCommands(sample_info_file,regions_file,\ 
features_file,cX=0,Rbatch_folder="",cor_matrix="",min_samples=25, sample_means_file="")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sample_info_file}{ Tab-separated file auto-generated by SamplePrepInfo() function - required for the whole analysis }
  \item{regions_file}{ Probe regions file }
  \item{features_file}{ A subset of probe regions file with rowcount, Chr regions, GC, deltaG and Tm as columns }
  \item{cX}{ [optional] cX=0 (default) excludes Chr X, cX=1 includes it }
  \item{Rbatch_folder}{ [optional] Rbatch folder bundled with CoNVex for batch execution. Rbatch_folder='' (default, works mostly) assumes that it's in R's library folder. You may copy it to somewhere else and modify this explicitly }
  \item{cor_matrix=""}{ [optional] Correlation matrix file (correlation between samples) default = "" }
  \item{min_samples=25}{ Minimum #samples expected in the subset. If chrX=1, both #males and #females should independently have this many number of samples }
  \item{sample_means_file}{ Default="" (don't save sample means) File to save sample means - this is useful for sample QC later. }
}
\details{
SampleLogRatio function calculates log2 ratio (of depth/sample-median) for each probe region. 
SampleLogRatioCall calls it for all samples. SampleLogRatioCallCommands (this function) generates Unix commands for batch execution.
}
\author{ 
Parthiban Vijayarangakannan\cr
Wellcome Trust Sanger Institute\cr
Cambridge, UK 
}
\note{  This can be easily used together with batch queuing (e.g., bsub) programs. This function usually generates a single Unix command for all samples in the analysis batch.
}
\seealso{ 
SampleLogRatio, SampleLogRatioCall
}
\examples{
> slr_command = SampleLogRatioCallCommands(sample_info_file,regions_file,\ 
features_file,cX=0,Rbatch_folder="");
> slr_command[1]
[1] "R --vanilla --slave --args '/home/pv1/MOPD_SampleInfo.txt, \ 
/home/pv1/R/x86_64-linux/2.11/CoNVex/extdata/SureSelect_50Mb.txt, \ 
/home/pv1/UK10K/UK10K_Features.txt,1' < \ 
/home/pv1/R/x86_64-linux/2.11/CoNVex/Rbatch/SampleLogRatioCall.R"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ log2 ratio }
